Code cleanup

+++ src/main/java/mod/schnappdragon/habitat/common/item/HabitatBoatItem.java
@@ -1,19 +1,19 @@
 package mod.schnappdragon.habitat.common.item;
 
 import mod.schnappdragon.habitat.common.entity.vehicle.HabitatBoat;
+import net.minecraft.stats.Stats;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntitySelector;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.stats.Stats;
-import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.entity.EntitySelector;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.phys.AABB;
 import net.minecraft.world.level.ClipContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
-import net.minecraft.world.level.Level;
 
 import java.util.List;
 import java.util.function.Predicate;
@@ -35,12 +35,12 @@ public class HabitatBoatItem extends Item {
         else {
             Vec3 vector3d = playerIn.getViewVector(1.0F);
             double d0 = 5.0D;
-                List<Entity> list = worldIn.getEntities(playerIn, playerIn.getBoundingBox().expandTowards(vector3d.scale(5.0D)).inflate(1.0D), ENTITY_PREDICATE);
+            List<Entity> list = worldIn.getEntities(playerIn, playerIn.getBoundingBox().expandTowards(vector3d.scale(5.0D)).inflate(1.0D), ENTITY_PREDICATE);
             if (!list.isEmpty()) {
                 Vec3 vector3d1 = playerIn.getEyePosition(1.0F);
 
-                for(Entity entity : list) {
-                    AABB axisalignedbb = entity.getBoundingBox().inflate((double)entity.getPickRadius());
+                for (Entity entity : list) {
+                    AABB axisalignedbb = entity.getBoundingBox().inflate(entity.getPickRadius());
                     if (axisalignedbb.contains(vector3d1))
                         return InteractionResultHolder.pass(itemstack);
                 }
@@ -62,8 +62,7 @@ public class HabitatBoatItem extends Item {
                     playerIn.awardStat(Stats.ITEM_USED.get(this));
                     return InteractionResultHolder.sidedSuccess(itemstack, worldIn.isClientSide());
                 }
-            }
-            else
+            } else
                 return InteractionResultHolder.pass(itemstack);
         }
     }
